refactor: import the sidebar grouping from the file that owns it
The split that moved the sidebar's grouping out of sessions.ts left a re-export
block behind, so every caller still reached the moved names through sessions.ts
and the file stayed the one import site for both concerns. That is the half of
the cohesion the split was for.

The block is gone and the six callers name sidebar-groups directly. Mechanical:
no export renamed, no behaviour touched, tsc and the suite green either way.
