+++ frontend/src/components/sidebar/SidebarRail.tsx
@@ -2,7 +2,8 @@ import { useMatch, useNavigate } from "react-router-dom"
 import { PanelLeft, Plus } from "lucide-react"
 import { cn } from "@/lib/utils"
 import { useProjects } from "@/providers/projects"
-import { activeSessionId, sessionsOf, sidebarGroups, type Session } from "@/lib/session/sessions"
+import { activeSessionId, sessionsOf, type Session } from "@/lib/session/sessions"
+import { sidebarGroups } from "@/lib/session/sidebar-groups"
 import { resolveGroups } from "@/lib/session/panes"
 import { useStoredGroups } from "@/lib/session/panes-store"
 import { useSessionAgent } from "@/lib/session/use-session-agent"
