+++ frontend/src/lib/session/sessions.test.ts
@@ -1,44 +1,46 @@
 import { describe, expect, it } from "vitest"
 import {
-  forkableSession,
-  forkUnavailableReason,
   activeSessionId,
-  dragOrder,
   activeTarget,
   addSession,
   adoptSession,
   closeSession,
+  delegatesOf,
   dropClosedSession,
-  groupByWorktree,
+  forkableSession,
+  forkUnavailableReason,
   hasSession,
-  isLastWorktreeSession,
   isSessionKind,
-  neighborSessionId,
-  orderGroups,
-  PINNED_GROUP_KEY,
   projectOfSession,
-  ROOT_GROUP_KEY,
   removeProject,
   renameSession,
   reorderSessions,
-  reorderSubset,
   restoreSession,
   resumableSession,
-  runCardIn,
   sessionOrigin,
   sessionsOf,
   setActiveSession,
   setSessionEntrypoint,
-  setSessionPinned,
   setSessionMCPServers,
+  setSessionPinned,
   setSessionSandboxed,
   setSessionSchedule,
-  delegatesOf,
-  sidebarGroups,
   type Session,
   type SessionKind,
   type SessionState,
 } from "./sessions"
+import {
+  dragOrder,
+  groupByWorktree,
+  isLastWorktreeSession,
+  neighborSessionId,
+  orderGroups,
+  PINNED_GROUP_KEY,
+  reorderSubset,
+  ROOT_GROUP_KEY,
+  runCardIn,
+  sidebarGroups,
+} from "./sidebar-groups"
 
 const P = "project-1"
